Parses CT Images with their relevant contours
Functional at the moment,  but it might be a good idea to define a "slice" class that has access to the CT data as well as all the relevant contour data for each image that we train the AI on. Implementing these functions on a slice class would be clearer.