Speed ups, Bug fixes, Periodic training and saving
Created a reward stack to be indexed instead of recalculating at each stage. Fixed the alignment of reward stack and binary stack indices.
Changed corresponding reset(), as well as removing lingering print statements and slight changes for clarity.
All (x,y) indexes were swapped for (y,x) indexing to match convention (and the way arrays are indexed).
Added in code to train and then save models in periods of 500000.